Add proper serialization

+++ src/types.rs
@@ -1,6 +1,6 @@
 /// Type alias for a judgement (see [`Statement`](../statement/struct.Statement.html))
-pub type Judgement = u16;
+pub type Judgement = u8;
 
-/// Type alias for a identifier representing a variable or operator (see
+/// Type alias for an identifier representing a variable or operator (see
 /// [`Expression`](../expression/struct.Expression.html))
 pub type Identifier = i16;
